pyproject.toml
[project]
name = "apiym"
version = "0.1.0"
description = "姚鸣的个人API服务"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "colorlog>=6.10.1",
    "fastapi>=0.128.0",
    "fastapi-mcp>=0.4.0",
    "jwt>=1.4.0",
    "nameko>=2.14.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
    "jinja2>=3.1.6",
    "bcrypt>=5.0.0",
    "httpx>=0.28.1",
    "pycryptodome>=3.23.0",
]
